# apps/backend/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "backend API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/admin": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_admins",
        "responses": {
          "200": {
            "description": "List all admins",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "add_admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Admin added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          },
          "409": {
            "description": "Admin already exists"
          }
        }
      }
    },
...
    "/api/admin/{twitch_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "operationId": "remove_admin",
        "parameters": [
          {
            "name": "twitch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Admin removed"
          },
          "403": {
            "description": "Cannot remove the last root admin"
          },
          "404": {
            "description": "Admin not found"
          },
          "409": {
            "description": "Cannot remove your own account"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddAdminRequest": {
        "type": "object",
        "required": [
          "twitch_id"
        ],
        "properties": {
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "twitch_id": {
            "type": "string"
          }
        }
      },
      "AdminResponse": {
        "type": "object",
        "required": [
          "twitch_id",
          "is_root",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "display_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_root": {
            "type": "boolean"
          },
          "twitch_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}